// src/common/enums/trip-status.enum.ts
export enum TripStatus {
  DRAFT = 'DRAFT',
  WAITING_FOR_DRIVER = 'WAITING_FOR_DRIVER',
  DRIVER_NOT_FOUND = 'DRIVER_NOT_FOUND',
  APPROVED = 'APPROVED',
  DRIVER_ON_WAY_TO_PICKUP = 'DRIVER_ON_WAY_TO_PICKUP',
  ARRIVED_AT_PICKUP = 'ARRIVED_AT_PICKUP',
  TRIP_IN_PROGRESS = 'TRIP_IN_PROGRESS',
  PAYMENT = 'PAYMENT',
  COMPLETED = 'COMPLETED',
  CANCELLED_PAYMENT = 'CANCELLED_PAYMENT',
  CANCELLED = 'CANCELLED',
}